refactor: move types to useMembers
